Software engineering/Exp6/ecommerce_recommender/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
bcrypt
email-validator
python-dotenv
Werkzeug
